refactor : applyservice 패키지명 apply로 변경
